[PF-48] Builder integration changes
- decouple library stage 1
